music/serializers.py
from rest_framework import serializers
from .models import Artist, Album, Song


class ArtistSerializer(serializers.ModelSerializer):

    class Meta:
        model = Artist
        fields = ("name", "image", "last_update")


class AlbumSerializer(serializers.ModelSerializer):
    artist = ArtistSerializer()

    class Meta:
        model = Album
        fields = ("title", "cover_image", "last_update", "artist")


class SongSerializer(serializers.ModelSerializer):
    album = AlbumSerializer()

    class Meta:
        model = Song
        fields = ("title", "cover_image", "last_update", "album")